Refactor receivers hierarchy (base abstract receiver)
