+++ libs/unity/library/Runtime/Scripts/Media/RemoteAudioRenderer.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
-guid: 43c74cfa42bccd246903f3459056188c
+guid: ecc506ef6d3213043bc0529b3f591d39
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
-  icon: {fileID: 2800000, guid: f845cb707dda5c84594956f49655100f, type: 3}
+  icon: {instanceID: 0}
   userData: 
   assetBundleName: 
   assetBundleVariant:
